Release TYPO3 14 animation improvements

+++ ext_tables.sql
@@ -3,10 +3,10 @@
 #
 CREATE TABLE tt_content (
   tx_content_gsap_animation_animation tinytext,
-  tx_content_gsap_animation_duration tinyint DEFAULT 800,
-  tx_content_gsap_animation_delay tinyint DEFAULT 0,
+  tx_content_gsap_animation_duration int(11) DEFAULT 800,
+  tx_content_gsap_animation_delay int(11) DEFAULT 0,
 
-  tx_content_gsap_animation_offset tinyint DEFAULT 0,
+  tx_content_gsap_animation_offset int(11) DEFAULT 0,
   tx_content_gsap_animation_once tinyint DEFAULT 1,
   tx_content_gsap_animation_mirror tinyint DEFAULT 0,
   tx_content_gsap_animation_easing tinytext,
